Developed the drop_duplicate and drop_nan
df.drop_duplicates() stopped doing the job as there started to be two different values for the very same observation. I also started dropping zero values, otherwise the drop_duplicates might have kept the observation with zero and not that with valid numbers.
Finally I improved the .sum() function to have it in line with the latest version